apps/ai-service/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
scikit-learn>=1.4.0
pandas>=2.2.0
pydantic>=2.6.1
supabase>=2.3.4
python-dotenv>=1.0.1
